pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "parsort"
version = "0.1.0"
description = "A human friendly CLI file sorter built around the PARA method."
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["pyyaml"]

[project.scripts]
parsort = "parsort.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]